move io/general docstrings
